# swagger.json
{
    "swagger":"2.0",
    "info":{
        "description":"This is a test server",
        "version":"1.0.0",
        "title":"KindergartenChiu"
    },
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths":{
...
        "/teacher/add":{
            "post":{
                "tags":["Teacher"],
                "parameters":[
                    {
                        "name":"Authorization",
                        "type":"string",
                        "in":"header",
                        "description":"Jwt sign by user's apiKey"
                    },
                    {
                        "name":"phone",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Teacher's phone"
                    },
                    {
                        "name":"password",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Teacher's password"
                    },
                    {
                        "name":"name",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Teacher's name"
                    }
                ],
                "responses":{
                    "200":{
                        "description": "OK",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "401":{
                        "decription": "Bad token",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "500":{
                        "decription": "Server fail",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    }
                }
            }
        },
        "/teacher/allTeacher":{
            "get":{
                "tags":["Teacher"],
                "parameters":[
                    {
                        "name":"Authorization",
                        "type":"string",
                        "in":"header",
                        "description":"Jwt sign by user's apiKey"
                    },
                    {
                        "name":"phone",
                        "type":"string",
                        "in":"Authorization",
                        "description":"user's phone"
                    }
                ],
                "responses":{
                    "200":{
                        "description": "OK",
                        "schema":{
                            "$ref": "#/definitions/Teacher"
                        }
                    },
                    "500":{
                        "decription": "Server fail",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    }
                }
            }
        },
...
        "/driver/add":{
            "post":{
                "tags":["Driver"],
                "parameters":[
                    {
                        "name":"Authorization",
                        "type":"string",
                        "in":"header",
                        "description":"Jwt sign by user's apiKey"
                    },
                    {
                        "name":"phone",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Driver's phone"
                    },
                    {
                        "name":"password",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Driver's password"
                    },
                    {
                        "name":"name",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Driver's name"
                    }
                ],
                "responses":{
                    "200":{
                        "description": "OK",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "401":{
                        "decription": "Bad token",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "500":{
                        "decription": "Server fail",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    }
                }
            }
        },
        "/driver/allDriver":{
            "get":{
                "tags":["Driver"],
                "parameters":[
                    {
                        "name":"Authorization",
                        "type":"string",
                        "in":"header",
                        "description":"Jwt sign by user's apiKey"
                    },
                    {
                        "name":"phone",
                        "type":"string",
                        "in":"Authorization",
                        "description":"user's phone"
                    }
                ],
                "responses": {
                    "200":{
                        "description": "OK",
                        "schema":{
                            "$ref": "#/definitions/Driver"
                        }
                    },
                    "500":{
                        "decription": "Server fail",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    }
                }
            }
        },
        "/driver/location":{
            "put":{
                "tags":["Driver"],
                "parameters":[
                    {
                        "name":"Authorization",
                        "type":"string",
                        "in":"header",
                        "description":"Jwt sign by user's apiKey"
                    },
                    {
                        "name":"phone",
                        "type":"string",
                        "in":"Authorization",
                        "description":"User's phone"
                    },
                    {
                        "name":"location",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Driver's location format: '<lat>,<lng>', example: '24.990750,121.294920' "
                    }
                ],
                "responses":{
                    "200":{
                        "description": "OK",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "401":{
                        "decription": "Bad token",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "500":{
                        "decription": "Server fail",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions":{
...
        "Teacher":{
            "type": "array",
            "items":{
                "type": "object",
                "properties":{
                    "uuid": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phone":{
                        "type": "string"
                    }
                }
            }
        },
        "Driver":{
            "type": "array",
            "items":{
                "type": "object",
                "properties":{
                    "uuid": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phone":{
                        "type": "string"
                    }
                }
            }
        },
        "message":{
            "type":"object",
            "properties":{
                "type":{
                    "type": "string"
                },
                "message":{
                    "type": "string"
                }
            }
        }
    }
}